refactor(checkable-input): move to internal folder
