+++ src/__experimental__/components/checkbox/checkbox.spec.js
@@ -3,11 +3,11 @@ import TestRenderer from "react-test-renderer";
 import { mount } from "enzyme";
 import { css } from "styled-components";
 import Checkbox from "./checkbox.component";
-import CheckableInput from "../checkable-input/checkable-input.component";
-import { StyledCheckableInput } from "../checkable-input/checkable-input.style";
+import CheckableInput from "../../../__internal__/checkable-input/checkable-input.component";
+import { StyledCheckableInput } from "../../../__internal__/checkable-input/checkable-input.style";
 import FieldHelpStyle from "../field-help/field-help.style";
-import HiddenCheckableInputStyle from "../checkable-input/hidden-checkable-input.style";
-import StyledCheckableInputSvgWrapper from "../checkable-input/checkable-input-svg-wrapper.style";
+import HiddenCheckableInputStyle from "../../../__internal__/checkable-input/hidden-checkable-input.style";
+import StyledCheckableInputSvgWrapper from "../../../__internal__/checkable-input/checkable-input-svg-wrapper.style";
 import guid from "../../../utils/helpers/guid";
 import { baseTheme } from "../../../style/themes";
 import {
