+++ src/__experimental__/components/checkbox/checkbox.style.js
@@ -2,10 +2,10 @@ import styled, { css } from "styled-components";
 import { space } from "styled-system";
 import PropTypes from "prop-types";
 import baseTheme from "../../../style/themes/base";
-import { StyledCheckableInput } from "../checkable-input/checkable-input.style";
+import { StyledCheckableInput } from "../../../__internal__/checkable-input/checkable-input.style";
 import StyledFieldHelp from "../field-help/field-help.style";
-import StyledHiddenCheckableInput from "../checkable-input/hidden-checkable-input.style";
-import StyledCheckableInputSvgWrapper from "../checkable-input/checkable-input-svg-wrapper.style";
+import StyledHiddenCheckableInput from "../../../__internal__/checkable-input/hidden-checkable-input.style";
+import StyledCheckableInputSvgWrapper from "../../../__internal__/checkable-input/checkable-input-svg-wrapper.style";
 import StyledLabel, { StyledLabelContainer } from "../label/label.style";
 import StyledValidationIcon from "../../../components/validations/validation-icon.style";
 import StyledFormField from "../form-field/form-field.style";
