+++ src/__experimental__/components/radio-button/radio-button-svg.component.js
@@ -1,5 +1,5 @@
 import React from "react";
-import StyledCheckableInputSvgWrapper from "../checkable-input/checkable-input-svg-wrapper.style";
+import StyledCheckableInputSvgWrapper from "../../../__internal__/checkable-input/checkable-input-svg-wrapper.style";
 
 const RadioButtonSvg = () => {
   return (
