+++ src/__experimental__/components/radio-button/radio-button.spec.js
@@ -4,9 +4,9 @@ import { ThemeProvider } from "styled-components";
 import TestRenderer from "react-test-renderer";
 import { RadioButton, RadioButtonGroup } from ".";
 import FieldHelpStyle from "../field-help/field-help.style";
-import HiddenCheckableInputStyle from "../checkable-input/hidden-checkable-input.style";
-import { StyledCheckableInput } from "../checkable-input/checkable-input.style";
-import StyledCheckableInputSvgWrapper from "../checkable-input/checkable-input-svg-wrapper.style";
+import HiddenCheckableInputStyle from "../../../__internal__/checkable-input/hidden-checkable-input.style";
+import { StyledCheckableInput } from "../../../__internal__/checkable-input/checkable-input.style";
+import StyledCheckableInputSvgWrapper from "../../../__internal__/checkable-input/checkable-input-svg-wrapper.style";
 import { assertStyleMatch } from "../../../__spec_helper__/test-utils";
 import guid from "../../../utils/helpers/guid";
 import baseTheme from "../../../style/themes/base";
