+++ src/__experimental__/components/switch/switch.style.js
@@ -2,9 +2,9 @@ import styled, { css } from "styled-components";
 import PropTypes from "prop-types";
 import baseTheme from "../../../style/themes/base";
 import FieldHelpStyle from "../field-help/field-help.style";
-import HiddenCheckableInputStyle from "../checkable-input/hidden-checkable-input.style";
+import HiddenCheckableInputStyle from "../../../__internal__/checkable-input/hidden-checkable-input.style";
 import { StyledLabelContainer } from "../label/label.style";
-import { StyledCheckableInput } from "../checkable-input/checkable-input.style";
+import { StyledCheckableInput } from "../../../__internal__/checkable-input/checkable-input.style";
 import StyledSwitchSlider from "./switch-slider.style";
 import StyledValidationIcon from "../../../components/validations/validation-icon.style";
 import { FieldLineStyle } from "../form-field/form-field.style";
